Sebas ramas

+++ src/main/java/com/usta/proyectointegrador/models/dao/TransactionDAO.java
@@ -7,13 +7,13 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface TransactionDAO extends JpaRepository<TransactionEntity, Integer> {
+public interface TransactionDAO extends JpaRepository<TransactionEntity, Long> {
 
     @Query("""
            SELECT t
            FROM TransactionEntity t
            WHERE t.nombreUsu.id_usuario = :usuarioId
            """)
-    List<TransactionEntity> findByUsuarioId(@Param("usuarioId") Integer usuarioId);
+    List<TransactionEntity> findByUsuarioId(@Param("usuarioId") Long usuarioId);
 
 }
